+++ apps/web-customer/src/components/blocks/HubBlocks/BillingSummaryBlock/BillingSummaryBlock.tsx
@@ -1,40 +1,8 @@
 import { Box } from "@mui/material";
-import React, { useEffect, useState } from "react";
-import { billlingSummaryMockData as data } from "../../../../core/constant/BillingSummaryMock/BillingSummaryMock"; // remove once real data fetching is applied
+import { billingSummaryColumns } from "../../../../core/constant/BillingSummaryMock/BillingSummaryMock";
 import { DataGrid } from "core-library/components";
-
 import { OrderSummaryBlock, OrderDetailsBlock } from "./index";
-
-const columns = [
-  {
-    field: "orderId",
-    sortable: false,
-    headerName: "Order ID",
-    flex: 1,
-    maxWidth: 450,
-  },
-  {
-    field: "dateTime",
-    sortable: false,
-    headerName: "Date/Time",
-    flex: 1,
-    maxWidth: 450,
-  },
-  {
-    field: "payment",
-    sortable: false,
-    headerName: "Payment",
-    flex: 1,
-    maxWidth: 450,
-  },
-  {
-    field: "purchase",
-    sortable: false,
-    headerName: "Purchase",
-    flex: 1,
-    maxWidth: 450,
-  },
-];
+import { useGetBillingSummary } from "../../../../core/hooks/useGetBillingSummary";
 
 const tableStyle = {
   border: "0.2px solid #B0BEC5",
@@ -65,25 +33,8 @@ const tableStyle = {
 };
 
 export const BillingSummaryBlock = () => {
-  // Uncomment this once the api has been created for fetching the billing summary:
-  // const { data, isLoading } = businessQueryGetBillingSummary(['getBillingSummary']);
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-  const [selectedRow, setSelectedRow] = useState<Record<string, any> | null>(
-    null
-  );
-
-  useEffect(() => {
-    if (data.length === 1) {
-      setSelectedRow(data[0]);
-    } else if (data.length > 1) {
-      const latestOrder = [...data].sort(
-        (a, b) =>
-          new Date(b.dateTime).getTime() - new Date(a.dateTime).getTime()
-      )[0];
-      setSelectedRow(latestOrder);
-    }
-  }, [data]);
-
+  const { isLoading, selectedData, setSelectedData, data } = useGetBillingSummary();
+  
   return (
     <section className="bg-hubBackground">
       <Box className="px-4 md:px-8 py-8 lg:py-16 w-full">
@@ -95,7 +46,7 @@ export const BillingSummaryBlock = () => {
             <Box className="p-6">
               <Box className="w-full max-w-[1100px] overflow-x-auto">
                 <DataGrid
-                  columns={columns.map((col) => ({
+                  columns={billingSummaryColumns.map((col) => ({
                     ...col,
                     flex: undefined,
                     width: col.flex ? 200 : undefined,
@@ -105,7 +56,7 @@ export const BillingSummaryBlock = () => {
                   initPageSize={5}
                   isLoading={isLoading}
                   autoHeight
-                  onRowClick={(params) => setSelectedRow(params.row)}
+                  onRowClick={(params) => setSelectedData(params.row)}
                   sx={tableStyle}
                 />
               </Box>
@@ -113,9 +64,9 @@ export const BillingSummaryBlock = () => {
           </Box>
 
           <Box className="flex flex-col md:flex-row lg:flex-col gap-4 w-auto lg:w-[323px]">
-            <OrderSummaryBlock selectedRow={selectedRow} />
+            <OrderSummaryBlock selectedRow={selectedData} />
 
-            <OrderDetailsBlock selectedRow={selectedRow} />
+            <OrderDetailsBlock selectedRow={selectedData} />
           </Box>
         </Box>
       </Box>
